+++ CHANGES.rst
@@ -4,12 +4,15 @@ Changelog
 0.5.6 (unreleased)
 ------------------
 
+- Add transform to change a href link on content-core and viewlet-below-content-body div.
+  [bsuttor]
+
 - Remove IMinisiteRoot marker interface on uninstall.
   [bsuttor]
 
 - 'sitemap.xml.gz' view is now callable from minisites.
   [bsuttor]
-
+ 
 
 0.5.5 (2018-06-06)
 ------------------